requirements.txt
annotated-doc==0.0.4 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.13.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2026.2.25 ; python_version >= "3.11" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.11" and python_version < "4.0" and platform_python_implementation != "PyPy"
charset-normalizer==3.4.6 ; python_version >= "3.11" and python_version < "4.0"
click==8.3.1 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
cryptography==46.0.6 ; python_version >= "3.11" and python_version < "4.0"
deprecated==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.135.2 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.49.1 ; python_version >= "3.11" and python_version < "4.0"
google-genai==1.70.0 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.3.2 ; python_version >= "3.11" and python_version < "4.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.16.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.11" and python_version < "4.0"
httptools==0.7.1 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.11 ; python_version >= "3.11" and python_version < "4.0"
jiter==0.13.0 ; python_version >= "3.11" and python_version < "4.0"
limits==5.8.0 ; python_version >= "3.11" and python_version < "4.0"
narwhals==2.18.1 ; python_version >= "3.11" and python_version < "4.0"
numpy==2.4.4 ; python_version >= "3.11" and python_version < "4.0"
ollama==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
openai==2.30.0 ; python_version >= "3.11" and python_version < "4.0"
packaging==26.0 ; python_version >= "3.11" and python_version < "4.0"
pandas==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
plotly==6.6.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.3 ; python_version >= "3.11" and python_version < "4.0"
pycparser==3.0 ; python_version >= "3.11" and python_version < "4.0" and implementation_name != "PyPy" and platform_python_implementation != "PyPy"
pydantic-core==2.41.5 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.12.5 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.2.2 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.3 ; python_version >= "3.11" and python_version < "4.0"
requests==2.33.1 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
slowapi==0.1.9 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.48 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.5.5 ; python_version >= "3.11" and python_version < "4.0"
starlette==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
tabulate==0.10.0 ; python_version >= "3.11" and python_version < "4.0"
tenacity==9.1.4 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.67.3 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2025.3 ; python_version >= "3.11" and python_version < "4.0" and (sys_platform == "win32" or sys_platform == "emscripten")
urllib3==2.6.3 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.42.0 ; python_version >= "3.11" and python_version < "4.0"
uvloop==0.22.1 ; python_version >= "3.11" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
vanna==2.0.2 ; python_version >= "3.11" and python_version < "4.0"
watchfiles==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
websockets==16.0 ; python_version >= "3.11" and python_version < "4.0"
wrapt==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
